move convert into new module
