+++ lenskit/lenskit/algorithms/knn/item.py
@@ -21,8 +21,7 @@ import torch
 from typing_extensions import Callable, Literal, Optional, TypeAlias, override
 
 from lenskit import util
-from lenskit.data import FeedbackType
-from lenskit.data.dataset import Dataset
+from lenskit.data import Dataset, FeedbackType
 from lenskit.data.vocab import Vocabulary
 from lenskit.diagnostics import ConfigWarning, DataWarning
 from lenskit.math.sparse import normalize_sparse_rows, safe_spmv
