+++ lenskit/lenskit/algorithms/knn/user.py
@@ -20,8 +20,7 @@ import torch
 from typing_extensions import NamedTuple, Optional, Self, override
 
 from lenskit import util
-from lenskit.data import FeedbackType
-from lenskit.data.dataset import Dataset
+from lenskit.data import Dataset, FeedbackType
 from lenskit.data.vocab import Vocabulary
 from lenskit.diagnostics import DataWarning
 from lenskit.math.sparse import normalize_sparse_rows, safe_spmv
