+++ lenskit/lenskit/algorithms/svd.py
@@ -12,7 +12,7 @@ import numpy as np
 import pandas as pd
 from typing_extensions import Literal, override
 
-from lenskit.data.dataset import Dataset
+from lenskit.data import Dataset
 from lenskit.data.vocab import Vocabulary
 
 try:
