+++ lenskit/tests/algorithms/test_als_explicit.py
@@ -16,8 +16,7 @@ from pytest import approx, mark
 import lenskit.util.test as lktu
 from lenskit import batch
 from lenskit.algorithms import als
-from lenskit.data.dataset import Dataset, from_interactions_df
-from lenskit.data.movielens import load_movielens_df
+from lenskit.data import from_interactions_df, load_movielens_df
 
 _log = logging.getLogger(__name__)
