+++ lenskit/tests/algorithms/test_candidate_selector.py
@@ -9,7 +9,7 @@ import pandas as pd
 
 import lenskit.util.test as lktu
 from lenskit.algorithms import basic
-from lenskit.data.dataset import Dataset, from_interactions_df
+from lenskit.data import Dataset, from_interactions_df
 
 simple_df = pd.DataFrame(
     {"item": [1, 1, 2, 3], "user": [10, 12, 10, 13], "rating": [4.0, 3.0, 5.0, 2.0]}
