+++ lenskit/tests/algorithms/test_svd.py
@@ -13,7 +13,7 @@ import pandas as pd
 from pytest import approx, mark
 
 from lenskit.algorithms import svd
-from lenskit.data.dataset import Dataset, from_interactions_df
+from lenskit.data import Dataset, from_interactions_df
 from lenskit.util import clone
 
 _log = logging.getLogger(__name__)
