+++ lenskit/tests/batch/test_batch_recommend.py
@@ -18,7 +18,7 @@ from lenskit import batch, topn
 from lenskit.algorithms import Recommender
 from lenskit.algorithms.basic import PopScore, TopN
 from lenskit.algorithms.bias import Bias
-from lenskit.data.dataset import Dataset, from_interactions_df
+from lenskit.data import Dataset, from_interactions_df
 
 _log = logging.getLogger(__name__)
