+++ lenskit/tests/eval/test_topn_analysis.py
@@ -17,7 +17,7 @@ from lenskit import crossfold as xf
 from lenskit.algorithms import Recommender
 from lenskit.algorithms.knn.item import ItemItem
 from lenskit.algorithms.knn.user import UserUser
-from lenskit.data.dataset import from_interactions_df
+from lenskit.data import from_interactions_df
 from lenskit.metrics.topn import _dcg, precision, recall
 from lenskit.util.test import demo_recs  # noqa: F401
