+++ lenskit/tests/splitting/test_split_users.py
@@ -13,7 +13,7 @@ import pandas as pd
 
 import pytest
 
-from lenskit.data.dataset import Dataset, from_interactions_df
+from lenskit.data import Dataset, from_interactions_df
 from lenskit.splitting.holdout import SampleFrac, SampleN
 from lenskit.splitting.users import crossfold_users, sample_users
